+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ### Enhancements
 - Similar errors on the same line will now not be removed
 - Render generic upper bound with `: ` instead of ` <: `
+- Render uninhabited type as `Never` instead of `<nothing>`
+- Render Callables with `-> None`
 ### Fixes
 - Handle positional only `/` parameters in overload implementation inference
 - Render inferred literal without `?`